requirements.txt
Django>=4.2.0
djangorestframework>=3.14.0
django-cors-headers>=4.0.0
neo4j>=5.0.0
openai>=1.0.0
python-dotenv>=1.0.0 